requirements.txt
streamlit
pandas
openpyxl

# Supabase Database
supabase>=2.0.0

# FastAPI Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0

# IRIS API Integration
requests>=2.28.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
cryptography>=41.0.0

# PDF Generation
reportlab>=4.0.0
pypdf2>=3.0.0

# Template Engine
jinja2>=3.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
